int -> int64_t

+++ hdr/convinience.h
@@ -1,7 +1,9 @@
-int max(int a, int b){
+#include <stdint.h>
+
+int64_t max(int64_t a, int64_t b){
     return a > b ? a : b;
 }
 
-int min(int a, int b){
+int64_t min(int64_t a, int64_t b){
     return a < b ? a : b;
 }
